Boilerplate code pretty much done.
